Expose namedQuery
